pyproject.toml
[tool.poetry]
name = "crypto-killer"
version = "0.1.0"
description = ""
authors = ["EraApp"]
readme = "README.md"

[tool.mypy]
python_version = "3.11:::"
ignore_missing_imports = true
warn_redundant_casts = true
disallow_untyped_defs = true
exclude = '(src/tests/)'


[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.pytest_cache
  | \.mypy_cache
  | \.venv
)/
'''

[tool.pycln]
all = true

[tool.bandit]
exclude_dirs = ["src/tests"]

[tool.isort]
profile = "black"
src_paths = ["src"]

[tool.coverage]
[tool.coverage.run]
omit = [
    "*/__init__.py",
    "*/constants.py",
]
[tool.coverage.report]
skip_empty = true

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto"
env = [
    "APP__ENV=TEST"
]

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.16"
asynchuobi = "^0.0.1"
beanie = "^1.17.0"
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
loguru = "^0.6.0"
ccxt = "^2.8.16"
python-telegram-bot = "13.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
